Update verbiage in channel_tutorial_tip
