+++ i18n/en.json
@@ -4168,8 +4168,13 @@
   "sidebar.sortedByRecencyLabel": "Recent activity",
   "sidebar.sortedManually": "Manually",
   "sidebar.team_select": "{siteName} - Join a team",
-  "sidebar.tutorialScreen1.body": "**Channels** organize conversations across different topics. They're open to everyone on your team. To send private communications use **Direct Messages** for a single person or **Private Channels** for multiple people.",
-  "sidebar.tutorialScreen1.title": "Channels",
+  "sidebar.tutorialChannelTypes.public": "Channels organize conversations across different topics. **Public channels** are open to everyone on your team.",
+  "sidebar.tutorialChannelTypes.private": "**Private channels** are for sensitive conversations between selected members.",
+  "sidebar.tutorialChannelTypes.direct": "**Direct messages** are for private conversations between a small group of up to seven people.",
+  "sidebar.tutorialChannelTypes.title": "Channels",
+
+  "//ASDF": "",
+
   "sidebar.tutorialScreen2.body1": "Here are two public channels to start:",
   "sidebar.tutorialScreen2.body2": "**{townsquare}** is a place for team-wide communication. Everyone in your team is a member of this channel.",
   "sidebar.tutorialScreen2.body3": "**{offtopic}** is a place for fun and humor outside of work-related channels. You and your team can decide what other channels to create.",
