+++ sass/components/_tutorial.scss
@@ -44,7 +44,7 @@
         }
 
         .arrow {
-            top: 80px;
+            top: 115px;
             left: -10px;
             border-right-color: $white;
             border-left-width: 0;
